style: components/interviewPage의 컴포넌트 구조변경
