+++ FE/src/components/interviewPage/InterviewFooter/RecordControlButton.tsx
@@ -1,0 +1,38 @@
+import { css } from '@emotion/react';
+import Icon from '@foundation/Icon/Icon';
+import Typography from '@foundation/Typography/Typography';
+
+type RecordControlButtonType = {
+  isRecording: boolean;
+  handleStartRecording: () => void;
+  handleStopRecording: () => void;
+};
+
+const RecordControlButton: React.FC<RecordControlButtonType> = ({
+  isRecording,
+  handleStartRecording,
+  handleStopRecording,
+}) => {
+  return (
+    <div
+      css={css`
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+        gap: 0.75rem;
+      `}
+      onClick={isRecording ? handleStopRecording : handleStartRecording}
+    >
+      {isRecording ? (
+        <Icon id="record-start" width="2rem" height="2rem" />
+      ) : (
+        <Icon id="record-start" width="2rem" height="2rem" />
+      )}
+      <Typography variant={'body1'} color="white">
+        {isRecording ? '녹화종료' : '녹화시작'}
+      </Typography>
+    </div>
+  );
+};
+export default RecordControlButton;
