+++ FE/src/components/interviewPage/InterviewHeader/InterviewHeader.tsx
@@ -1,7 +1,7 @@
 import { css } from '@emotion/react';
 import RecordStatus from './RecordStatus';
-import Typography from '../foundation/Typography/Typography';
 import RecordTimer from './RecordTimer';
+import IntervieweeName from './IntervieweeName';
 
 type InterviewHeaderProps = {
   isRecording: boolean;
@@ -26,22 +26,7 @@ const InterviewHeader: React.FC<InterviewHeaderProps> = ({
       `}
     >
       <RecordStatus isRecording={isRecording} />
-      <div
-        css={css`
-          max-width: 600px;
-          overflow: hidden;
-          text-overflow: ellipsis;
-          white-space: normal;
-
-          @media (max-width: 46rem) {
-            width: 9.375rem;
-          }
-        `}
-      >
-        <Typography noWrap paragraph variant={'body1'} color="white">
-          {intervieweeName}
-        </Typography>
-      </div>
+      <IntervieweeName intervieweeName={intervieweeName} />
       <RecordTimer isRecording={isRecording} />
     </div>
   );
