+++ FE/src/components/interviewPage/InterviewHeader/RecordTimer.tsx
